commits telas e funções de treino

+++ lib/domain/entities/workout/exercice_entity.dart
@@ -1,7 +1,6 @@
-import 'dart:ffi';
 
 class ExerciseEntity {
-  final Long id;
+  final int id;
   final String name;
   final int sets;
   final int reps;
